添加了新的项目在内，包括贪吃蛇和简易俄罗斯方块 之前的Movie MVVM，由于实在对于目前的学习没有帮助，反而打击信心，暂且放弃 1、学习第一行代码，过程中学习在内的项目，详细查看 2、学习开发的思路，以及独立开发的思维

+++ MovieMVVM/app/src/main/java/com/overz/moviemvvm/data/vo/MovieDetails.kt
@@ -8,11 +8,11 @@ data class MovieDetails(
     val id: Int,
     val overview: String,
     val popularity: Double,
-    @SerializedName("poster_path")      //告诉JSON这个和下面的两个是同一个东西，只是名字大小写
+    @SerializedName("poster_path")          //告诉JSON这个和下面的两个是同一个东西，只是名字大小写
     val posterPath: String,
     @SerializedName("release_date")
     val releaseDate: String,
-    val revenue: Int,
+    val revenue: Long,
     val runtime: Int,
     val status: String,
     val tagline: String,
